## Summary
- default the covenant signer HTTP listener to loopback and require a
bearer auth token for non-loopback exposure
- stop holding the submit mutex across engine.OnSubmit so deduped
callers are not serialized behind long signer work
- add focused tests for auth enforcement, config flags, and in-flight
submit dedupe semantics

## Testing
- go test ./pkg/covenantsigner -count=1
- go test ./pkg/tbtc -run 'TestCovenantSignerEngine_' -count=1
- go test ./cmd -count=1

+++ pkg/tbtc/covenant_signer.go
@@ -523,7 +523,7 @@ func (cse *covenantSignerEngine) buildAndSignSelfV1Transaction(
 	if err := builder.SetInputSequence(0, request.MigrationTransactionPlan.InputSequence); err != nil {
 		return nil, fmt.Errorf("cannot set covenant input sequence: %v", err)
 	}
-	builder.SetLocktime(uint32(request.MigrationTransactionPlan.LockTime))
+	builder.SetLocktime(request.MigrationTransactionPlan.LockTime)
 	builder.AddOutput(&bitcoin.TransactionOutput{
 		Value:           destinationValue,
 		PublicKeyScript: destinationScript,
@@ -614,7 +614,7 @@ func (cse *covenantSignerEngine) buildQcV1SignerHandoff(
 	if err := builder.SetInputSequence(0, request.MigrationTransactionPlan.InputSequence); err != nil {
 		return nil, fmt.Errorf("cannot set covenant input sequence: %v", err)
 	}
-	builder.SetLocktime(uint32(request.MigrationTransactionPlan.LockTime))
+	builder.SetLocktime(request.MigrationTransactionPlan.LockTime)
 	builder.AddOutput(&bitcoin.TransactionOutput{
 		Value:           destinationValue,
 		PublicKeyScript: destinationScript,
